Rename cxi.mpi_submit to cctbx.xfel.submit_job
