+++ xfel/util/mp.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, division, print_function
 #
 # Handle multiprocessing with any of the implemented methods so that this step
-# is abstracted away from the use case (e.g. cxi_mpi_submit).
+# is abstracted away from the use case (e.g. cctbx.xfel.submit_job).
 #
 from libtbx.utils import Sorry
 import os
